Search updated
In index.html - Changed car manufacturer to a text input instead of a datalist.

In app.js - Added more fields to the catalogue search. The user can now search by make, model, car number along with the catalogue number and the search narrow the results as the user types.
Removed function for obtaining carManufacturers as it is no longer required.