Changes to forms and server
When launching the server it now loads the index.html page which for current purposes allows user to fiew the catalogue.

Places all the elements in the middle of the page in a col-md-4 element.

Added hyperlinks to the other pages that are present.

Added an upload option on the add a car page, this will store the path of the image so it can be viewed in the view catalogue page.

Added an img element for the picture of the car to be placed.

Created a css file - removes bullet points from ul/centers the headers on the pages.

Added the formidable package, to deal with uploads.

The service history page contains a datalist similar to view catalogue page, and all the necessary identifying pieces of information elements, so the user can be sure they have selected the correct car.