Set image path for new cars
When adding a new car, the image path is set so that it can be retrieved and displayed when viewing the car via the catalogue.

Fixed spelling mistake in datalist and the Car model that was stopping results from returning correctly.

Added multer for file uploads(not required so will be removed).