Can add new services from the site
Can now create a new service record from the site and it links to a car accordingly.

Added a table to the catalogue page, which will display the different services a car has had.

Add a edit a car page which will use a put request to update a record in the database.